Replace community name template tag with Python community.
The tag sometimes refers to the Python community as a whole, and
sometimes refers to the group of Python leaders who will be enforcing
the Code of Conduct. I've replaced some tags with a reference to the
Python Code of Conduct team. (Please note that team does not exist yet.)