Add community norms from PSF Python Code of Conduct.
Move the three community norms from the original PSF Python Code of
Conduct into this new document.

I excluded this particular paragraph:

"Overall, we're good to each other. We contribute to this community not
because we have to, but because we want to. If we remember that, these
guidelines will come naturally."

This paragraph is problematic for several reasons.

"We're good to each other" is a sentence that downplays the sexism,
racism, transphobia, and ableism that many marginalized people in tech
face when they attempt to contribute to a technical community. A white
cis able-bodied person is more likely to receive goodwill from community
members. Saying that everyone is "good to each other" does not
acknowledge that people from groups under-represented in tech may have a
different experience and not experience goodwill.

Finally, the sentence itself is vague as to what being "good to each
other" means.

"We contribute to this community not because we have to, but because we
want to."

The underlying assumption in this sentence is that people contribute to
open source projects because they are passionate about it, because it's
a hobby. However, many marginalized people in tech are financially
unable to contribute to open source projects on their own time. They may
be able to participate in open source if it becomes a part of their paid
job.

We should not downplay people who submit good contributions because it's
part of their job requirements. Not everyone has the privilege of being
a passionate hobbyist.

https://www.ashedryden.com/blog/the-ethics-of-unpaid-labor-and-the-oss-community

"If we remember that, these guidelines will come naturally."

Removed this sentence because the previous two sentences were removed.