Map Contributor Covenant cultural norms to sentences in original Python Code of Conduct
It turns out that the Contributor Covenant's concise language maps to
different sentences in the original Python Code of Conduct.

The only sentence that was removed was "Members of the community are
respectful." That was redundant with the heading "Being respectful".

I moved sentences referring to different kinds of respect into more
specific categories. Respect for people's time is now under a new
"Acknowledging time and effort" category. Respecting community processes
is under the "Focusing on what is best for the community" category from
the Contributor Covenant.

Not sure if these are quite in the right order, and I'm open to
suggestions.