Add section on reversisms from LGBTQ in Tech Slack Code of Conduct.
These statements are very important for a community adopting a new Code
of Conduct. They clarify that "tone policing" is often used to silence
people who experience marginalization and discrimination on a daily
basis.

People who face discrimination or harassment are not required to be nice
to people who have privilege. They are not required to educate them. A
Code of Conduct is in place so that discrimination and harassment stops.
It should not be weaponized against people who face discrimination and
harassment.

Additionally, this section covers boundary pushing behavior. Often
times, people with personality disorders or who are grooming people for
sexual harassment will object to people setting boundaries. They will
attempt to find a way to punish a person who is setting a boundary. It's
important to recognize that people have the right to set a boundary.  If
the other person does not accept that boundary, they have the right to
reiterate that boundary in stronger and stronger terms.

There is still plenty of room for nuance in this interpretation to
evaluate interpersonal communication issues. The Code of Conduct team
should still intervene to mitigate harmful and disrespectful
communication where ever possible. However, they are under no obligation
to respond to a reversism report.