Add encouraging self-harm to list of inappropriate behaviors.
This language was taken from the LGBTQ in Tech Slack Code of Conduct:

https://lgbtq.technology/coc.html

It's important to include this sentence because "threats of violence"
may not include encouraging others to commit self-harm or suicide.
Before Twitter changed their rules to include this inappropriate
behavior, they said that a person tweeting a gif of someone jumping out
the window in response to someone being trans gender was not against
their terms of service. This covers that case.

This also covers the case of someone getting others to harass an
individual. We see this sometimes with people who gather up friends on
discussion boards to harass a specific person.