Update PSF enforcement procedures with edits from CoC WG committee members.
Replace Americanism "toe the line" with "borderline inappropriate
behavior" and fix typo.

Clarify that there is a conflict of interest if the reporter or reported
person is a metamour of a Code of Conduct committee member. If they are
a metamour of someone in general, there's no conflict of interest.

Emphasize that incidents that happen outside the community where a
community member's mental health or physical safety would be impacted
may be in scope.

Emphasize that private conversations in community spaces are also in
scope. This was mentioned in the impact section, but not the
jurisdiction section.